feat: move manager models to models file
